requirements.txt
# Core dependencies
streamlit>=1.32.0
python-docx>=1.1.0
Pillow>=10.0.0
huggingface_hub>=0.21.0

# OpenAI for GPT-4 Vision
openai>=1.0.0
